requirements.txt
Django>=5.2.1
djangorestframework>=3.16.0
django-channels>=0.7.0
psycopg2-binary>=2.9.10
redis>=6.2.0
python-dotenv>=1.1.0
channels-redis>=4.2.0
django-cors-headers>=4.3.1
Pillow>=10.2.0  # For image handling
gunicorn>=21.2.0  # For production deployment
uvicorn>=0.27.1  # For ASGI server 